code robustness, improved error handling in buzzer task

+++ main/Buzzer.h
@@ -33,8 +33,8 @@ public:
 	static void init( uint frequency=2700 );
 	static void volume( uint vol=0 );
 	static void frequency( uint f);
-	static void play2( uint f1=BUZZ_DH, uint d1=200, uint v1=100, uint f2=BUZZ_E, uint d2=200, uint v2=0, uint repetition=1 );
-	static void play( uint freq=BUZZ_DH, uint duration=200, uint volume=100, uint freq2=BUZZ_DH, uint duration2=0, uint volume2=100 );
+	static void play2( uint16_t f1=BUZZ_DH, uint16_t d1=200, uint16_t v1=100, uint16_t f2=BUZZ_E, uint16_t d2=200, uint16_t v2=0, uint repetition=1 );
+	static void play( uint16_t freq=BUZZ_DH, uint16_t duration=200, uint16_t volume=100, uint16_t freq2=BUZZ_DH, uint16_t duration2=0, uint16_t volume2=100 );
 private:
 	static TaskHandle_t pid;
 	static void taskStart();
